LayoutTests/mathml/presentation/sup-nested.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
</head>
<body>
<p>This test passes if it does not time out.</p>
<math>
  <msup>
    <mi>a</mi>
    <msup>
      <mi>b</mi>
      <msup>
        <mi>c</mi>
        <msup>
          <mi>d</mi>
          <msup>
            <mi>e</mi>
            <msup>
              <mi>f</mi>
              <mo>g</mo>
            </msup>
          </msup>
        </msup>
      </msup>
    </msup>
  </msup>
</math>
</body>
</html>
